- migrate to raspberry

+++ app/src/main/java/com/example/skolapp/VolleyCallBack.java
@@ -1,6 +1,6 @@
 package com.example.skolapp;
 
 public interface VolleyCallBack{
-    void onSuccess(float value);
+    void onSuccess(double value);
 
 }
